Main

Documentation

alphabet :: [Char]

Checks the first and last characters of a string for punctuation and removes them

clean :: [String] -> [String]

makeLower :: [String] -> [String]

Maps to lower across a list of strings (frequencies are case-insensitive)

toMap :: [String] -> Map String Int

Passes in an empty map to be populated

trans :: [String] -> Map String Int -> Map String Int

Wrapper function for addCount

addCount :: [String] -> Map String Int -> Map String Int

Adds occurrences of words to an immutable map

printMap :: Map String Int -> String

Builds string output from map, calculates proper spacing

getMaxKey :: Map String Int -> Int

Helper function to retrieve the maximum key (largest word) in the map

sorter :: [Char] -> [Char] -> Ordering

Output should be in descending order by highest frequency

main :: IO ()

Main method wrapper

toTry :: IO ()

Main method body

handler :: IOError -> IO ()

Error handler